% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doingpak-package.R
\docType{package}
\name{doingpak-package}
\alias{doingpak}
\alias{doingpak-package}
\title{doingpak: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patysosacosta/doingpak}
  \item Report bugs at \url{https://github.com/patysosacosta/doingpak/issues}
}

}
\author{
\strong{Maintainer}: Patricia Sosa \email{patsa@kth.se}

Authors:
\itemize{
  \item Patricia Sosa \email{patsa@kth.se}
}

}
\keyword{internal}
